+++ packages/cli/src/config/defineConfig.js
@@ -1,6 +1,0 @@
-/**
- * 用于智能提示
- */
-export function defineConfig(config) {
-    return config;
-}