+++ packages/cli/src/config/env.js
@@ -1,7 +1,0 @@
-/**
- * check dev env
- * @returns boolean true is dev
- */
-export function isDev() {
-    return process.env.NODE_ENV === 'development';
-}