+++ packages/vue-components/src/components/index.ts
@@ -8,21 +8,7 @@ import ChatContent from './RTASR/ChatContent.vue';
 import Recorder from './RTASR/Recorder.vue';
 import UniWebview from './UniWebview/index.vue';
 import Uploader from './Uploader/index.vue';
-import DragResize from './DragResize/index.vue';
 import Tags from './Tags/index.vue';
-export {
-  AlgoClient,
-  VLottiePlayer,
-  MP4BoxVideoPlayer,
-  VideoPlayer,
-  Realtime,
-  JsonForm,
-  ChatContent,
-  Recorder,
-  UniWebview,
-  Uploader,
-  DragResize,
-  Tags,
-};
+export { AlgoClient, VLottiePlayer, MP4BoxVideoPlayer, VideoPlayer, Realtime, JsonForm, ChatContent, Recorder, UniWebview, Uploader, Tags };
 
 export * from './Icon';